Rename files to drop the 'builder'-prefix
