+++ src/dapla_pseudo/types.py
@@ -8,7 +8,7 @@ import fsspec
 import gcsfs
 import pandas as pd
 
-from dapla_pseudo.v1.models import Field
+from dapla_pseudo.v1.api_models import Field
 
 FieldDecl = str | dict[str, str] | Field
 BinaryFileDecl = t.Union[
