+++ src/dapla_pseudo/v1/validation.py
@@ -11,8 +11,8 @@ import polars as pl
 import requests
 
 from dapla_pseudo.utils import get_file_format_from_file_name
-from dapla_pseudo.v1.builder_models import Result
 from dapla_pseudo.v1.ops import _client
+from dapla_pseudo.v1.result import Result
 from dapla_pseudo.v1.supported_file_format import read_to_polars_df
